update sept 2018

+++ libft/ft_strncmp.c
@@ -6,15 +6,15 @@
 /*   By: shcohen <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/05/17 21:50:03 by shcohen           #+#    #+#             */
-/*   Updated: 2018/05/23 19:36:37 by shcohen          ###   ########.fr       */
+/*   Updated: 2018/06/19 17:42:22 by shcohen          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int		ft_strncmp(const char *s1, const char *s2, unsigned int n)
+int		ft_strncmp(const char *s1, const char *s2, size_t n)
 {
-	unsigned int	i;
+	size_t	i;
 
 	i = 0;
 	if (n == 0)
